+++ core/src/ten_manager/designer_frontend/src/components/Popup/Editor.tsx
