+++ core/src/ten_manager/designer_frontend/src/components/Popup/Global.tsx
@@ -12,10 +12,10 @@ import {
   EWidgetCategory,
   EWidgetDisplayType,
 } from "@/types/widgets";
-import TerminalPopup from "@/components/Popup/TerminalPopup";
-import EditorPopup from "@/components/Popup/EditorPopup";
-import CustomNodeConnPopup from "@/components/Popup/CustomNodeConnPopup";
-import { LogViewerPopup } from "@/components/Popup/LogViewerPopup";
+import TerminalPopup from "@/components/Popup/Terminal";
+import EditorPopup from "@/components/Popup/Editor";
+import CustomNodeConnPopup from "@/components/Popup/CustomNodeConn";
+import { LogViewerPopup } from "@/components/Popup/LogViewer";
 import { GraphSelectPopup } from "@/components/Popup/Default/GraphSelect";
 import { AboutPopup } from "@/components/Popup/Default/About";
 import { PreferencesPopup } from "@/components/Popup/Default/Preferences";
@@ -30,7 +30,7 @@ import {
   ExtensionPopup,
 } from "@/components/Popup/Default/Extension";
 import { DocRefPopup } from "@/components/Popup/Default/DocRef";
-import { GraphPopup } from "@/components/Popup/GraphPopup";
+import { GraphPopup } from "@/components/Popup/Graph";
 import { groupWidgetsById } from "@/components/Popup/utils";
 
 export function GlobalPopups() {
