+++ core/src/ten_manager/designer_frontend/src/flow/FlowCanvas.tsx
@@ -25,7 +25,7 @@ import CustomNode from "@/flow/CustomNode";
 import CustomEdge from "@/flow/CustomEdge";
 import NodeContextMenu from "@/flow/ContextMenu/NodeContextMenu";
 import EdgeContextMenu from "@/flow/ContextMenu/EdgeContextMenu";
-import { TerminalData } from "@/components/Popup/TerminalPopup";
+import { TerminalData } from "@/components/Popup/Terminal";
 import { ThemeProviderContext } from "@/components/theme-context";
 import { cn } from "@/lib/utils";
 import { useWidgetStore, useAppStore } from "@/store";
